requirements.txt
alpaca-trade-api
python-dotenv
pandas
numpy
streamlit
plotly
streamlit-autorefresh
requests
